TUF targets modified.
Our targets are pairs consisting of git repositories and commits.
Not actual target files. So, generation of targets metadata (delegated roles to be exact)
was changed. Also updated targets format.
Additionally, the delegated targets metadata can now also contain custom information. We need this in order to specify capstone (if a commit is the last commit on a branch).