Refactored updater to support different types of targets and ways of downloading data
Our use case has two main differences compared to the current implementation of the updater. Firstly, we have a git repository which contains the metadata files. We do not want to download these files from a remote server. The users should have local clones of these repositories. So, downloading newer metadata in our case means fetching new commits and reading the committed metadata. We do not need to have previous and current directories. We can just checkout different commits. Also, we can just read contents of a file using git show. We do not need to read data from physical files.

The second thing is that in our case targets are not files, but commits in git repositories. So, updating targets means updating git repositories. Validation needs to ensure that history of a repository was not modified by an attacker. We need to verify that the commits are in accordance with the metadata. We do not need to check hashes or lengths of individual files.

What this refacotring focuses on is separating validation of metadata (and operations like traversing target roles) from loading and downloading these files. Once the metadata is loaded/updated and is stored in memory, the same checks can be executed, regardless of how the first two actions were carried out. In order to provide the needed flexibility, all code for loading and downloading metadata was moved to a separate class. Just like there is an implementation of that class which loads metadata from the file system and downloads these files from a server, there can be another implementation where git repositories are used. These classes are referred to as metadata handlers. When instantiating the updater, one only needs to specify which class of the metadata handler should be used. Similarly, everything related to downloading targets and validating if the downloaded files are valid was moved to another class. This classes is reffered to as the targets handler.

None of the existing functionalities of the updater were modified. The updater's tests should all still be passing. The only difference is that, due to some of the methods being moved to the handler, the tests need to call the metadata/target handler's methods.