Work toward making the process of updating metadata files a bit more flexible
These changes are nececssary for defining a different way of updating metadata, such
as by loading the files from a cloned git repository. These other ways should all be able
to use the concept of current and previous repositories, but might not want to download
the files from repote locations. If we want to implement the git updater,
we could think of 'file_mirrors' in  '_get_metadata_file' as git commits. So, we could
iterate over these commits. So, we want to redefine specification of the mirrors. Similarly,
we do not want to download a file from a remote location, but to read it from the git repository
at a specific revision. In case of the git updater, we also want to keep track of the current
commit, so added a 'on_successful_update' method. Also, we might not want to show the same
error message if the update ends up being unseccessful.