Work on the git updater
This is an attempt at getting the git updater working using
the concept of current and previous directories with as
little refactoring as possible. The updater reads content
of a file from the cloned bare git repository (using git show).
That content is then used to create a temp file which is returned.
Git commits are treated as mirrors.
This appears to be working, but there are a couple of things that
should be different in case of the git updater, so there will
be additional modifications.