Using kwargs in the handlers
We do not always need the same date in our git handler as
the original tuf implementation needs. By using kwargs, we
can avoid having to add pylint disable. This was necessary
in order to get the build passing. Removed long docstrings as
they just made the code less readable and this is never going
to get merged upstream in any case